Created sensors function
