+++ VentilatorControl/src/sensors.cpp
@@ -1,0 +1,47 @@
+#include "includes.h"
+
+double getPressureReadings (void)
+{
+  if(micros() - PressureSensorLastStatusRead >= 1000)
+  {
+    I2CMux.openChannel(PATIENT_CIRCUIT_PRESSURE_MUX_CHANNEL);
+    uint8_t sensorStatus = patientCircuitPressure.readStatus();
+    if( !patientCircuitPressure.isBusy(sensorStatus) && !patientCircuitPressure.isError(sensorStatus) )
+    {  
+      if(!patientCircuitPressure.readData(false))
+      {
+        // Get a pressure reading and convert to units of cmH2O
+        if( patientCircuitPressure.pressure > 0 )
+          pressure_reading = patientCircuitPressure.pressure * INCHES_2_CM;
+        patientCircuitPressure.startMeasurement();
+      }
+    }
+    I2CMux.closeChannel(PATIENT_CIRCUIT_PRESSURE_MUX_CHANNEL);
+
+    I2CMux.openChannel(VENTRUI_DIFFERENTIAL_PRESSURE_MUX_CHANNEL);
+    sensorStatus = venturiDifferentialPressure.readStatus();
+    if( !venturiDifferentialPressure.isBusy(sensorStatus) && !venturiDifferentialPressure.isError(sensorStatus) )
+    {  
+      if(!venturiDifferentialPressure.readData(false))
+      {
+        // Get a pressure reading and convert to units of cmH2O
+        if( venturiDifferentialPressure.pressure > 0 )
+        {
+          venturiDifferentialPressureReading = venturiDifferentialPressure.pressure * INCHES_2_CM;
+          // Transfer function found experimentally through bench testing
+          // Flow [lpm] = f(pressure [cmH20])
+          // see here for data:
+          // https://docs.google.com/spreadsheets/d/1bI_WWhnsqxKvRou-n7BQC0ML1uTgSl8-lZcnprczsNk/edit?usp=sharing
+          venturiFlowRateLpm = 26.486*sqrt(venturiDifferentialPressureReading)-0.276;
+          
+          // lowest reported flow rate = 0lpm
+          venturiFlowRateLpm = venturiFlowRateLpm>0?venturiFlowRateLpm:0;
+        }
+        venturiDifferentialPressure.startMeasurement();
+      }
+    }
+    I2CMux.closeChannel(VENTRUI_DIFFERENTIAL_PRESSURE_MUX_CHANNEL);
+    PressureSensorLastStatusRead = micros();
+  }
+  return pressure_reading;
+}
