+++ VentilatorControl/src/sensors.h
@@ -1,0 +1,6 @@
+#ifndef __SENSORS_H__
+#define __SENSORS_H__
+
+double getPressureReadings (void);
+
+#endif // __SENSORS_H__
